requirements.txt
colour-science[optional]


